Refactored Android audio to be more easily extendable and replaceable
