+++ backends/gdx-backend-android/src/com/badlogic/gdx/backends/android/AndroidApplication.java
@@ -119,7 +119,7 @@ public class AndroidApplication extends Activity implements AndroidApplicationBa
 		graphics = new AndroidGraphics(this, config, config.resolutionStrategy == null ? new FillResolutionStrategy()
 			: config.resolutionStrategy);
 		input = AndroidInputFactory.newAndroidInput(this, this, graphics.view, config);
-		audio = new AndroidAudio(this, config);
+		audio = new DefaultAndroidAudio(this, config);
 		this.getFilesDir(); // workaround for Android bug #10515463
 		files = new AndroidFiles(this.getAssets(), this.getFilesDir().getAbsolutePath());
 		net = new AndroidNet(this, config);
@@ -495,5 +495,10 @@ public class AndroidApplication extends Activity implements AndroidApplicationBa
 	@Override
 	public Handler getHandler () {
 		return this.handler;
+	}
+
+	@Override
+	public AndroidAudio createAudio (Context context, AndroidApplicationConfiguration config) {
+		return new DefaultAndroidAudio(context, config);
 	}
 }
